🔍 Update types to include SEO
