+++ types/Articles.ts
@@ -1,12 +1,36 @@
-import { Article } from './Project'
+import { ImageMetadata } from 'sanity'
+import { MetaProps } from './Meta'
+
+export type ArticleItems = {
+  _id: string
+  _createdAt: Date
+  name: string
+  slug: string
+  image: {
+    url: string
+    metadata: ImageMetadata
+  }
+  url: string
+  summary: string
+  duration: string
+}
+
+export interface Article {
+  seo: MetaProps
+  items: ArticleItems[]
+}
 
 interface ArticlesProps {
-  articles: Article[]
+  articles: ArticleItems[]
+  seo: MetaProps
 }
 
 interface FeaturedArticlesInterface {
   link: string
-  imgSrc: string
+  imgSrc: {
+    url: string
+    metadata: ImageMetadata
+  }
   title: string
   duration: string
   summary: string
